% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentPolish.R
\name{CurrentPolish}
\alias{CurrentPolish}
\title{Read and polish raw current}
\usage{
CurrentPolish(file, TimeStart, TimeEnd, voltage = 50, MaxCurrent = 130)
}
\arguments{
\item{file}{The path of .ABF file.}

\item{TimeStart, TimeEnd}{The start and end of signal time (s) of ABF file.}

\item{voltage}{The expement voltage.}
}
\description{
Read and polish raw current
}
