% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtract.R
\name{FeatureExtract}
\alias{FeatureExtract}
\title{Normalize and extract features for machine learning}
\usage{
FeatureExtract(x)
}
\arguments{
\item{x}{Identified amino acids signal, see also \code{SignalExtract}}
}
\description{
Normalize and extract features for machine learning
}
