% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{Internal}
\alias{Internal}
\alias{KBins2}
\alias{mean2}
\alias{sd2}
\alias{BUB_Sig}
\title{Internal}
\usage{
KBins2(sig, minseglen1 = 100, minseglen2 = 10, pen.value = 0.001)

mean2(x, q1 = 0.25, q2 = 0.75)

sd2(x, q1 = 0.1, q2 = 0.9)

BUB_Sig(Mat, LB = 2000)
}
\description{
Internal

Internal

Internal

Internal
}
\concept{Internal}
