% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LevelIdentify.R
\name{LevelIdentify}
\alias{LevelIdentify}
\title{Signal level identify}
\usage{
LevelIdentify(object, L0Min = NA, L0Max = NA, L1Min = NA, L1Max = NA)
}
\arguments{
\item{object}{The result of function `CurrentPolish`.}

\item{L0Min, L0Max, L1Min, L1ax}{The default range of L0 and L1 signal level.}
}
\description{
Signal level identify
}
