% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{Predict}
\alias{Predict}
\title{Predict the amino acids of signal}
\usage{
Predict(x, model)
}
\arguments{
\item{x}{Features of signal.}

\item{model}{Trained model.}
}
\description{
Predict the amino acids of signal
}
