% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalExtract.R
\name{SignalExtract}
\alias{SignalExtract}
\title{Extract signal from polished current}
\usage{
SignalExtract(
  object,
  L0Min = NA,
  L0Max = NA,
  L1Min = NA,
  L1Max = NA,
  MinDwellTime = 0.00075
)
}
\arguments{
\item{object}{The result of function `CurrentPolish`.}

\item{L0Min, L0Max, L1Min, L1ax}{The default range of L0 and L1 signal level.}

\item{MinDwellTime}{The minimum dwell time for signal filtering.}
}
\description{
Extract signal from polished current
}
